requirements.txt
altair==4.2.0
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
astroid==2.3.3
asttokens==2.0.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.6.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
boto3==1.20.28
botocore==1.23.28
Brotli==1.0.9
cachetools==5.0.0
catalogue==2.0.6
certifi==2019.11.28
cffi==1.15.0
cfn-flip==1.3.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==3.4.8
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
durationpy==0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
fasteners==0.14.1
filelock==3.4.0
Flask==2.0.2
fonttools==4.28.5
future==0.18.2
gilfoyle==0.915
gitdb==4.0.9
GitPython==3.1.25
glob2==0.7
googletrans==4.0.0rc1
gpg===1.13.1-unknown
h11==0.9.0
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.0.1
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
idle==1.0.4
idna==2.8
importlib-resources==5.4.0
imutils==0.5.4
ipykernel==6.6.1
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.4
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
JPype1==1.3.0
jsonschema==4.3.2
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaitaistruct==0.9
kappa==0.6.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.0.1
konlpy==0.5.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.9.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.1.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mecab-python===0.996-ko-0.9.2
mistune==0.8.4
mitmproxy==7.0.4
monotonic==1.5
msgpack==1.0.3
murmurhash==1.0.6
mypy==0.761
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
packaging==21.3
pafy==0.5.5
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
passlib==1.7.4
pathy==0.6.1
pep517==0.12.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip-tools==6.4.0
Pivy==0.6.5
placebo==0.10.0
platformdirs==2.4.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.18.1
psutil==5.5.1
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-python-adb==0.3.0.dev0
pyarrow==6.0.1
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
pydeck==0.7.1
pydyf==0.1.2
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyphen==0.12.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
pyserial==3.4
PySocks==1.6.8
pysrt==1.1.2
pytesseract==0.3.8
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-opensubtitles==0.6.dev0
python-slugify==5.0.2
python-vlc==3.0.12118
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
roman==2.0.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scipy==1.7.3
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.1.0
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
streamlit==1.3.1
systemd-python==234
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.13
thonny==3.2.7
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
torbrowser-launcher==0.3.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
troposphere==3.2.1
tweepy==4.4.0
typed-ast==1.4.1
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.8
urwid==2.1.2
usb-creator==0.3.7
validators==0.18.2
virtualenv==20.10.0
wadllib==1.3.3
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
weasyprint==53.4
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.11.2
wsgi-request-logger==0.4.6
wsproto==1.0.0
xkit==0.0.0
zappa==0.54.1
zipp==3.6.0
zopfli==0.1.9
zstandard==0.15.2
